[FIX] migrated 06_heartbeats.js to use the new event name as per jetstream migration docs [CHANGE] changed the examples to be plain javascript, as this makes it easier to use out of the box.
