+++ jetstream/examples/01_consumers.js
@@ -15,30 +15,21 @@
 
 import { connect } from "@nats-io/transport-deno";
 import { jetstream } from "@nats-io/jetstream";
-import { setupStreamAndConsumer } from "./util.ts";
+import { setupStreamAndConsumer } from "./util.js";
 
 // create a connection
 const nc = await connect();
 
-// create a stream with a random name with some messages and a consumer
 const { stream, consumer } = await setupStreamAndConsumer(nc);
 
 // retrieve an existing consumer
 const js = jetstream(nc);
 const c = await js.consumers.get(stream, consumer);
+console.log(await c.info(true));
 
-// the consumer is wrapped in loop because this way, if there's some failure
-// it will re-setup consume, and carry on.
-// this is the basic pattern for processing messages forever
-while (true) {
-  console.log("waiting for messages");
-  const messages = await c.consume();
-  try {
-    for await (const m of messages) {
-      console.log(m.seq);
-      m.ack();
-    }
-  } catch (err) {
-    console.log(`consume failed: ${(err as Error).message}`);
-  }
-}
+// getting an ordered consumer requires no name
+// as the library will create it
+const oc = await js.consumers.get(stream);
+console.log(await oc.info(true));
+
+await nc.close();
