+++ jetstream/tests/jstest_util.ts
@@ -83,3 +83,15 @@ export function fill(
 
   return Promise.all(a);
 }
+
+export async function setupStreamAndConsumer(
+  nc: NatsConnection,
+  messages = 100,
+): Promise<{ stream: string; consumer: string }> {
+  const stream = nuid.next();
+  await initStream(nc, stream, { subjects: [`${stream}.*`] });
+  await fill(nc, stream, messages, { randomize: true });
+  const consumer = await createConsumer(nc, stream);
+
+  return { stream, consumer };
+}
